Using generic type for TxStatus, reorganizing files
