+++ packages/api/src/chaintypes/substrate/tx.ts
@@ -1,99 +1,101 @@
 // Generated by @dedot/codegen
+
 import type {
-  MultiAddressLike,
-  Extrinsic,
-  BytesLike,
   AccountId32Like,
-  Percent,
-  Perbill,
-  H256,
-  FixedU128,
+  BytesLike,
   Data,
+  Extrinsic,
   FixedBytes,
+  FixedU128,
   FixedU64,
+  H256,
+  MultiAddressLike,
+  Perbill,
+  Percent,
   Perquintill,
+  TransactionStatus,
 } from '@dedot/codecs';
 import type {
   GenericChainTx,
   GenericTxCall,
+  IRuntimeTxCall,
   ISubmittableExtrinsic,
   ISubmittableResult,
-  IRuntimeTxCall,
 } from '@dedot/types';
 import type {
-  KitchensinkRuntimeRuntimeCallLike,
-  SpRuntimeMultiSignature,
+  FrameSupportPreimagesBounded,
+  FrameSupportScheduleDispatchTime,
   FrameSystemEventRecord,
   KitchensinkRuntimeOriginCaller,
-  SpWeightsWeightV2Weight,
-  SpConsensusSlotsEquivocationProof,
-  SpSessionMembershipProof,
-  SpConsensusBabeDigestsNextConfigDescriptor,
-  PalletElectionProviderMultiPhaseRawSolution,
-  PalletElectionProviderMultiPhaseSolutionOrSnapshotSize,
-  SpNposElectionsElectionScore,
-  SpNposElectionsSupport,
-  PalletStakingRewardDestination,
-  PalletStakingValidatorPrefs,
-  PalletStakingPalletConfigOp,
-  PalletStakingPalletConfigOpU32,
-  PalletStakingPalletConfigOpPercent,
-  PalletStakingPalletConfigOpPerbill,
-  KitchensinkRuntimeSessionKeys,
-  FrameSupportPreimagesBounded,
-  PalletDemocracyVoteAccountVote,
-  PalletDemocracyConviction,
-  PalletDemocracyMetadataOwner,
-  PalletElectionsPhragmenRenouncing,
-  SpConsensusGrandpaEquivocationProof,
-  PalletContractsWasmDeterminism,
-  PalletImOnlineHeartbeat,
-  PalletImOnlineSr25519AppSr25519Signature,
-  PalletIdentityIdentityInfo,
-  PalletIdentityBitFlags,
-  PalletIdentityJudgement,
-  PalletVestingVestingInfo,
   KitchensinkRuntimeProxyType,
-  PalletMultisigTimepoint,
-  PalletUniquesDestroyWitness,
-  PalletNftsCollectionConfig,
-  PalletNftsDestroyWitness,
-  PalletNftsMintWitness,
-  PalletNftsItemConfig,
-  PalletNftsBitFlags,
-  PalletNftsAttributeNamespace,
-  PalletNftsCancelAttributesApprovalWitness,
-  PalletNftsMintSettings,
-  PalletNftsItemTip,
-  PalletNftsPriceWithDirection,
-  PalletNftsPreSignedMint,
-  PalletNftsPreSignedAttributes,
+  KitchensinkRuntimeRuntimeCallLike,
+  KitchensinkRuntimeSessionKeys,
+  PalletAllianceCid,
+  PalletAllianceDisbandWitness,
+  PalletAllianceUnscrupulousItem,
+  PalletAssetConversionNativeOrAssetId,
+  PalletBrokerConfigRecord,
+  PalletBrokerCoreMask,
+  PalletBrokerFinality,
+  PalletBrokerRegionId,
+  PalletBrokerScheduleItem,
+  PalletContractsWasmDeterminism,
+  PalletConvictionVotingConviction,
+  PalletConvictionVotingVoteAccountVote,
   PalletCoreFellowshipParamsType,
   PalletCoreFellowshipWish,
-  SpTransactionStorageProofTransactionStorageProof,
+  PalletDemocracyConviction,
+  PalletDemocracyMetadataOwner,
+  PalletDemocracyVoteAccountVote,
+  PalletElectionProviderMultiPhaseRawSolution,
+  PalletElectionProviderMultiPhaseSolutionOrSnapshotSize,
+  PalletElectionsPhragmenRenouncing,
+  PalletIdentityBitFlags,
+  PalletIdentityIdentityInfo,
+  PalletIdentityJudgement,
+  PalletImOnlineHeartbeat,
+  PalletImOnlineSr25519AppSr25519Signature,
+  PalletMultisigTimepoint,
+  PalletNftsAttributeNamespace,
+  PalletNftsBitFlags,
+  PalletNftsCancelAttributesApprovalWitness,
+  PalletNftsCollectionConfig,
+  PalletNftsDestroyWitness,
+  PalletNftsItemConfig,
+  PalletNftsItemTip,
+  PalletNftsMintSettings,
+  PalletNftsMintWitness,
+  PalletNftsPreSignedAttributes,
+  PalletNftsPreSignedMint,
+  PalletNftsPriceWithDirection,
+  PalletNominationPoolsBondExtra,
+  PalletNominationPoolsClaimPermission,
+  PalletNominationPoolsCommissionChangeRate,
+  PalletNominationPoolsConfigOp,
+  PalletNominationPoolsConfigOp004,
+  PalletNominationPoolsConfigOpPerbill,
+  PalletNominationPoolsConfigOpU32,
+  PalletNominationPoolsPoolState,
+  PalletStakingPalletConfigOp,
+  PalletStakingPalletConfigOpPerbill,
+  PalletStakingPalletConfigOpPercent,
+  PalletStakingPalletConfigOpU32,
+  PalletStakingRewardDestination,
+  PalletStakingValidatorPrefs,
   PalletStateTrieMigrationMigrationLimits,
   PalletStateTrieMigrationMigrationTask,
   PalletStateTrieMigrationProgress,
-  FrameSupportScheduleDispatchTime,
-  PalletConvictionVotingVoteAccountVote,
-  PalletConvictionVotingConviction,
-  PalletAllianceDisbandWitness,
-  PalletAllianceCid,
-  PalletAllianceUnscrupulousItem,
-  PalletNominationPoolsBondExtra,
-  PalletNominationPoolsPoolState,
-  PalletNominationPoolsConfigOp,
-  PalletNominationPoolsConfigOpU32,
-  PalletNominationPoolsConfigOpPerbill,
-  PalletNominationPoolsConfigOp004,
-  PalletNominationPoolsClaimPermission,
-  PalletNominationPoolsCommissionChangeRate,
-  PalletAssetConversionNativeOrAssetId,
-  PalletBrokerConfigRecord,
-  PalletBrokerScheduleItem,
-  PalletBrokerRegionId,
-  PalletBrokerCoreMask,
-  PalletBrokerFinality,
+  PalletUniquesDestroyWitness,
+  PalletVestingVestingInfo,
+  SpConsensusBabeDigestsNextConfigDescriptor,
+  SpConsensusGrandpaEquivocationProof,
+  SpConsensusSlotsEquivocationProof,
+  SpNposElectionsElectionScore,
+  SpNposElectionsSupport,
+  SpRuntimeMultiSignature,
+  SpSessionMembershipProof,
+  SpTransactionStorageProofTransactionStorageProof,
+  SpWeightsWeightV2Weight,
 } from './types.js';
 
 export type ChainSubmittableExtrinsic<T extends IRuntimeTxCall = KitchensinkRuntimeRuntimeCallLike> = Extrinsic<
@@ -102,7 +104,7 @@ export type ChainSubmittableExtrinsic<T extends IRuntimeTxCall = KitchensinkRunt
   SpRuntimeMultiSignature,
   any[]
 > &
-  ISubmittableExtrinsic<ISubmittableResult<FrameSystemEventRecord>>;
+  ISubmittableExtrinsic<ISubmittableResult<FrameSystemEventRecord, TransactionStatus>>;
 
 export type TxCall = (...args: any[]) => ChainSubmittableExtrinsic;
