+++ packages/api/src/extrinsic/submittable/index.ts
@@ -1,0 +1,2 @@
+export * from './SubmittableResult.js';
+export * from './SubmittableExtrinsic.js';
